NES APU complete overhaul (bug fixes, refactors, working DMC)
  * Implemented NESDL_APU::Reset to reset some values between games
  * Modified APU variable types, casting to fix issues, removed unused frame counter code
  * Removed some quirks, bugs
    * APU cpuClocksPerSample was decreased slightly to avoid pops, this doesn't seem to be an issue. Removed this
    * Fixed some janky sequencer timing counting, we now perform better timing & cleaner logic
    * Removed bizarre DMC rate division by 4 (was trying to get the "correct" sound before, code was just buggy)
    * Casting values to bigger types during register writes, since bit-shifting was taking place
    * Fixed DMC bytes left counter not being assigned correctly (this was causing SO MUCH headache, why DMC sounded off)
  * Implemented APU IRQ (have not tested fully, I have seen it fire but haven't tested its accuracy)
  * Moved chunk of code to its own function (UpdateAPUFrameCounter)

APU is in a much better state but not perfect, I'm not a big fan of how it sounds but the fact that it's working great is very pleasing. May follow up with some accuracy validation fixes, or move on to other things, as audio is a low priority outside of necessary things (the basics, IRQ and CPU quirks, etc). I just really wanted to get the DMC channel working, it bothered me so much how wrong it sounded